ref: add contours and methods modules in kde submodule
